+++ src/components/admin/index.js
@@ -1,0 +1,33 @@
+import AdminHeader from '@/components/admin/ui/AdminHeader'
+import AdminTable from '@/components/admin/ui/AdminTable'
+import AssetsAdmin from '@/components/admin/AssetsAdmin'
+import ChannelsAdmin from '@/components/admin/ChannelsAdmin'
+import CollectionsAdmin from '@/components/admin/CollectionsAdmin'
+import DashboardAdmin from '@/components/admin/DashboardAdmin'
+import FormsAdmin from '@/components/admin/FormsAdmin'
+import LanguagesAdmin from '@/components/admin/LanguagesAdmin'
+import LayoutAdmin from '@/components/admin/ui/LayoutAdmin'
+import PagesAdmin from '@/components/admin/PagesAdmin'
+import SettingsAdmin from '@/components/admin/SettingsAdmin'
+import SideMenu from '@/components/admin/ui/SideMenu'
+import SignIn from '@/components/admin/login/SignIn'
+import SignUp from '@/components/admin/login/SignUp'
+import UsersAdmin from '@/components/admin/UsersAdmin'
+
+export {
+  AdminHeader,
+  AdminTable,
+  AssetsAdmin,
+  ChannelsAdmin,
+  CollectionsAdmin,
+  DashboardAdmin,
+  FormsAdmin,
+  LanguagesAdmin,
+  LayoutAdmin,
+  PagesAdmin,
+  SettingsAdmin,
+  SideMenu,
+  SignIn,
+  SignUp,
+  UsersAdmin
+}